[Toast] Refactor to useRenderElement
