+++ packages/react/src/toast/action/ToastAction.tsx
@@ -1,10 +1,9 @@
 'use client';
 import * as React from 'react';
 import type { BaseUIComponentProps } from '../../utils/types';
-import { useComponentRenderer } from '../../utils/useComponentRenderer';
 import { useToastRootContext } from '../root/ToastRootContext';
-import { mergeProps } from '../../merge-props';
 import { useButton } from '../../use-button/useButton';
+import { useRenderElement } from '../../utils/useRenderElement';
 
 /**
  * Performs an action when clicked.
@@ -13,14 +12,14 @@ import { useButton } from '../../use-button/useButton';
  * Documentation: [Base UI Toast](https://base-ui.com/react/components/toast)
  */
 export const ToastAction = React.forwardRef(function ToastAction(
-  props: ToastAction.Props,
+  componentProps: ToastAction.Props,
   forwardedRef: React.ForwardedRef<HTMLButtonElement>,
 ) {
-  const { render, className, disabled, ...other } = props;
+  const { render, className, disabled, ...elementProps } = componentProps;
 
   const { toast } = useToastRootContext();
 
-  const computedChildren = toast.actionProps?.children ?? other.children;
+  const computedChildren = toast.actionProps?.children ?? elementProps.children;
   const shouldRender = Boolean(computedChildren);
 
   const { getButtonProps } = useButton({
@@ -35,14 +34,17 @@ export const ToastAction = React.forwardRef(function ToastAction(
     [toast.type],
   );
 
-  const { renderElement } = useComponentRenderer({
-    render: render ?? 'button',
+  const renderElement = useRenderElement('button', componentProps, {
     ref: forwardedRef,
-    className,
     state,
-    extraProps: mergeProps<'button'>(toast.actionProps, other, getButtonProps, {
-      children: computedChildren,
-    }),
+    props: [
+      elementProps,
+      toast.actionProps,
+      getButtonProps,
+      {
+        children: computedChildren,
+      },
+    ],
   });
 
   if (!shouldRender) {
